Added new HtlpHelper
