+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/SFA.DAS.ProviderApprenticeshipsService.Web.csproj
@@ -594,7 +594,7 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Helpers\FeatureEnabledCheck.cs" />
+    <Compile Include="Helpers\HtmlExtensions.cs" />
     <Compile Include="Models\AcknowledgementViewModel.cs" />
     <Compile Include="Models\Agreement\AgreementsViewModel.cs" />
     <Compile Include="Models\Agreement\CommitmentAgreement.cs" />
